+++ app/src/Navigation/MainNavigator.tsx
@@ -1,12 +1,12 @@
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import React, {PropsWithChildren} from 'react';
-import ForgotPasswordScreen from '../Components/Screens/Locked/ForgotPassword';
-import LogInScreen from '../Components/Screens/Locked/LogIn';
-import ResetPasswordScreen from '../Components/Screens/Locked/ResetPassword';
-import SignUpScreen from '../Components/Screens/Locked/SignUp';
+import ForgotPasswordScreen from '../Components/Screens/Welcome/ForgotPassword';
+import LogInScreen from '../Components/Screens/Welcome/LogIn';
+import ResetPasswordScreen from '../Components/Screens/Welcome/ResetPassword';
+import SignUpScreen from '../Components/Screens/Welcome/SignUp';
 import SplashScreen from '../Components/SplashScreen';
 import WelcomeScreen from '../Components/Screens/Welcome/Welcome';
-import VerifyEmailScreen from '../Components/Screens/Locked/VerifyEmail';
+import VerifyEmailScreen from '../Components/Screens/Welcome/VerifyEmail';
 import AuthContext from '../Context/authContext';
 import {AuthContextType} from '../types/AuthContextType';
 import {WelcomeNavigationRoutesType} from '../types/NavigationRoutesType';
